requirements.txt
langchain_community==0.4.1
langchain_core==1.2.5
langchain_google_genai==4.1.2
langchain_openai==1.1.6
python-dotenv==1.2.1
streamlit==1.52.2
